refactor to break apart resources and start adding tests
