+++ render/preprocessor/control/case_test.go
@@ -63,47 +63,3 @@ task.query "query" {
 		assert.Equal(t, "macro.case", caseNode.Kind())
 	})
 }
-
-// // TestCaseIsTrue ensures truthiness
-// func TestCaseIsTrue(t *testing.T) {
-//	c := &control.Case{}
-//	t.Run("true when predicate is 'true'", func(t *testing.T) {
-//		c.Predicate = "true"
-//		actual, err := c.IsTrue()
-//		assert.NoError(t, err)
-//		assert.True(t, actual)
-//	})
-//	t.Run("false when predicate is 'false'", func(t *testing.T) {
-//		c.Predicate = "false"
-//		actual, err := c.IsTrue()
-//		assert.NoError(t, err)
-//		assert.False(t, actual)
-//	})
-//	t.Run("true when predicate is true template expression", func(t *testing.T) {
-//		c.Predicate = "eq 1 1"
-//		actual, err := c.IsTrue()
-//		assert.NoError(t, err)
-//		assert.True(t, actual)
-//	})
-//	t.Run("false when predicate is false template expression", func(t *testing.T) {
-//		c.Predicate = "eq 1 0"
-//		actual, err := c.IsTrue()
-//		assert.NoError(t, err)
-//		assert.False(t, actual)
-//	})
-//	t.Run("error when predicate is empty", func(t *testing.T) {
-//		c.Predicate = ""
-//		_, err := c.IsTrue()
-//		assert.Error(t, err)
-//	})
-//	t.Run("error when predicate is non-truthy value", func(t *testing.T) {
-//		c.Predicate = "sven"
-//		_, err := c.IsTrue()
-//		assert.Error(t, err)
-//	})
-//	t.Run("error when template returns error", func(t *testing.T) {
-//		c.Predicate = "eq 1"
-//		_, err := c.IsTrue()
-//		assert.Error(t, err)
-//	})
-// }
