+++ render/preprocessor/control/resources_test.go
@@ -25,8 +25,16 @@ import (
 // TestInterfacesAreImplemented ensures that all types implement the correct
 // interfaces
 func TestInterfacesAreImplemented(t *testing.T) {
-	t.Run("SwitchPreparer", func(t *testing.T) { assert.Implements(t, (*resource.Resource)(nil), new(control.SwitchPreparer)) })
-	t.Run("SwitchTask", func(t *testing.T) { assert.Implements(t, (*resource.Task)(nil), new(control.SwitchTask)) })
-	t.Run("CasePreparer", func(t *testing.T) { assert.Implements(t, (*resource.Resource)(nil), new(control.CasePreparer)) })
-	t.Run("CaseTask", func(t *testing.T) { assert.Implements(t, (*resource.Task)(nil), new(control.CaseTask)) })
+	t.Run("SwitchPreparer", func(t *testing.T) {
+		assert.Implements(t, (*resource.Resource)(nil), new(control.SwitchPreparer))
+	})
+	t.Run("SwitchTask", func(t *testing.T) {
+		assert.Implements(t, (*resource.Task)(nil), new(control.SwitchTask))
+	})
+	t.Run("CasePreparer", func(t *testing.T) {
+		assert.Implements(t, (*resource.Resource)(nil), new(control.CasePreparer))
+	})
+	t.Run("CaseTask", func(t *testing.T) {
+		assert.Implements(t, (*resource.Task)(nil), new(control.CaseTask))
+	})
 }
